requirements.txt
slack-bolt
python-dotenv
aiohttp
notion-client
apscheduler
pytz
google-generativeai
httpx
